requirements.txt
alabaster==0.7.12
autopep8==1.5.2
Babel==2.9.1
certifi==2023.7.22
coverage==5.1
dateutils==0.6.6
defusedxml==0.6.0
django-autocomplete-light==3.3.4
django-autoslug==1.9.4
django-bootstrap3==11.0.0
django-bower==5.1.0
django-ckeditor==5.7.1
django-crontab==0.7.1
django-chunked-upload==2.0.0
django-debug-toolbar==3.2.1
django-extensions==2.1.6
django-filter==2.4.0
django-formtools==2.1
django-fsm==2.4.0
django-js-asset==1.2.2
django-modeltranslation==0.13.1
django-mptt==0.10.0
django-multiselectfield==0.1.8
django-multiupload==0.5.2
django-ordered-model==3.3.0
django-recaptcha==1.3.1
django-reversion==3.0.2
django-solo==2.0.0
django-tables2==2.0.0a3
Django==2.2.28
djangorestframework==3.11.2
docutils==0.16
elasticsearch==2.3.0
elasticstack==0.5.0
et-xmlfile==1.0.1
chardet==3.0.4
idna==3.7
imagesize==1.2.0
jdcal==1.4.1
Jinja2==3.1.3
odfpy==1.4.1
odictliteral==1.0.0
openpyxl==3.1.5
pandas==1.3.5
Pillow==9.5.0
psycopg2==2.8.6
pycodestyle==2.5.0
Pygments==2.15.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-memcached==1.58
pytz==2020.1
PyYAML==5.4
raven==6.10.0
requests==2.31.0
six==1.14.0
snowballstemmer==2.0.0
sorl-thumbnail==12.8.0
Sphinx==1.4.4
sqlparse==0.4.4
tablib[all]==3.4.0
tld==0.13
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==4.1.1
urllib3==1.26.18
xlrd==1.2.0
xlwt==1.3.0
playwright==1.48.0
